moved the chat server to its own separate module
