+++ chat-server/index.js
@@ -1,18 +1,18 @@
-const express = require('express');
+const express = require('../server/node_modules/express');
 const CHATPORT = process.env.CHATPORT || 8080
 const socketIo = require('socket.io');
 const cors = require('cors');
-const chatApp = express();
+const app = express();
 const NetAddress = require('./utils/network');
 const chat = require('./utils/chat-handler')
-const server = require('http').createServer(chatApp);
+const server = require('http').createServer(app);
 const io = socketIo(server, {
     cors: {
         origin: '*',
         methods: ['GET', 'POST']
     }
 })
-chatApp.use(cors())
+app.use(cors())
 
 const ChatServer = () => {
     server.listen(CHATPORT, () => {
