+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node server/server.js",
     "develop": "concurrently \"cd server && npm run watch \" \"cd client && npm start\"",
-    "install": "cd server && npm i && cd ../client && npm i",
+    "install": "cd server && npm i && cd ../client && npm i && cd ../chat-server && npm i ",
     "seed": "cd server && npm run seed",
     "build": "cd client && npm run build"
   },
