+++ server/package.json
@@ -23,9 +23,7 @@
     "graphql": "^15.5.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.9.9",
-    "request-ip": "^2.1.3",
-    "socket.io": "^4.1.3",
-    "uuid": "^8.3.2"
+    "request-ip": "^2.1.3"
   },
   "devDependencies": {
     "nodemon": "^2.0.3"
