+++ server/server.js
@@ -3,7 +3,7 @@ const app = express();
 const path = require('path');
 const PORT = process.env.PORT || 3001;
 const db = require('./config/connection');
-const ChatServer = require('./chatServer');
+const ChatServer = require('../chat-server');
 const { ApolloServer } = require('apollo-server-express');
 const { authMiddleware } = require('./utils/auth');
 const requestIp = require('request-ip');
@@ -45,6 +45,6 @@ db.once('open', () => {
     console.log(`API server running on port ${PORT}!`);
     // log where we can go to test our GQL API
     console.log(`Use GraphQL at http://localhost:${PORT}${server.graphqlPath}`);
+    ChatServer()
   });
-  ChatServer();
 });
